Move the global shortcuts and window file import out of App.vue
The app-wide keydown handler and the window-level drag/drop/paste import
each become a composable that registers and tears down its own listener,
so App.vue's onMounted stops being a list of five window listeners. Both
still reach the sidebar and grid through the template refs they are
handed.

App.vue drops from 1735 to 1515 lines.
